Tell Reporters which spectra have been copied, and separate absolutely everything GUI related into desktop subpackage
