--- src/mora_the_explorer/explorer/__init__.py
@@ -1,7 +1,0 @@
-# This import is crucial (both that it happens and that it happens before the others) as
-# without a running QCoreApplication instance all the threading and signals and so on
-# will not work
-from .app import App
-
-# These are just namespace imports for convenience
-from .explorer import Explorer
